requirements.txt
pycryptodomex==3.9.4
urllib3==1.25.7
prometheus_client
